+++ lib/src/multi_flag_manager.dart
@@ -1,4 +1,5 @@
 import 'package:file_state_manager/file_state_manager.dart';
+import 'package:collection/collection.dart';
 
 ///
 /// This package manages multiple flag lists identified by name and supports serialization and deserialization.
@@ -9,7 +10,7 @@ import 'package:file_state_manager/file_state_manager.dart';
 ///
 class MultiFlagManager extends CloneableFile {
   static const String className = 'MultiFlagManager';
-  static const String version = '3';
+  static const String version = '4';
   final Map<String, List<bool>> _map = {};
   static const String _saveKey = 'map';
 
@@ -101,5 +102,29 @@ class MultiFlagManager extends CloneableFile {
   /// 通常はこれを直接呼び出さないでください。
   Map<String, List<bool>> getMap() {
     return _map;
+  }
+
+  @override
+  bool operator ==(Object other) {
+    if (other is MultiFlagManager) {
+      if (_map.length != other._map.length) {
+        return false;
+      }
+      for (String key in _map.keys) {
+        if (!other._map.containsKey(key) ||
+            !const ListEquality().equals(_map[key], other._map[key])) {
+          return false;
+        }
+      }
+      return true;
+    } else {
+      return false;
+    }
+  }
+
+  @override
+  int get hashCode {
+    List<Object> objects = [UtilObjectHash.calcMap(_map)];
+    return Object.hashAll(objects);
   }
 }