+++ lib/src/multi_index_manager.dart
@@ -1,3 +1,4 @@
+import 'package:collection/collection.dart';
 import 'package:file_state_manager/file_state_manager.dart';
 
 ///
@@ -101,5 +102,29 @@ class MultiIndexManager extends CloneableFile {
   /// 通常はこれを直接呼び出さないでください。
   Map<String, Set<int>> getMap() {
     return _map;
+  }
+
+  @override
+  bool operator ==(Object other) {
+    if (other is MultiIndexManager) {
+      if (_map.length != other._map.length) {
+        return false;
+      }
+      for (String key in _map.keys) {
+        if (!other._map.containsKey(key) ||
+            !const SetEquality().equals(_map[key], other._map[key])) {
+          return false;
+        }
+      }
+      return true;
+    } else {
+      return false;
+    }
+  }
+
+  @override
+  int get hashCode {
+    List<Object> objects = [UtilObjectHash.calcMap(_map)];
+    return Object.hashAll(objects);
   }
 }