+++ lib/src/textfield_manager.dart
@@ -10,7 +10,7 @@ import 'package:flutter/material.dart';
 ///
 class TextFieldManager extends CloneableFile {
   static const String className = 'TextFieldManager';
-  static const String version = '7';
+  static const String version = '8';
   final Map<String, TextEditingController> _ctrlMap = {};
   final Map<String, FocusNode> _focusMap = {};
   static const String _oldSaveKey = 'text_map';
@@ -184,5 +184,32 @@ class TextFieldManager extends CloneableFile {
       }
     }
     return false;
+  }
+
+  /// (en) Only TextEditingControllers are considered for comparison.
+  ///
+  /// (ja) TextEditingControllers のみが比較対象になります。
+  @override
+  bool operator ==(Object other) {
+    if (other is TextFieldManager) {
+      if (_ctrlMap.length != other._ctrlMap.length) {
+        return false;
+      }
+      for (String key in _ctrlMap.keys) {
+        if (!other._ctrlMap.containsKey(key) ||
+            _ctrlMap[key]!.text != other._ctrlMap[key]!.text) {
+          return false;
+        }
+      }
+      return true;
+    } else {
+      return false;
+    }
+  }
+
+  @override
+  int get hashCode {
+    List<Object> objects = [UtilObjectHash.calcMappedTEC(_ctrlMap)];
+    return Object.hashAll(objects);
   }
 }